Create README.md
I am pleased to present my latest project, which involves automating work with the Chrome web browser to test end-to-end (E2E) scenarios. As the test object, I have chosen a website that focuses on helping and saving the lives of dogs in the UK. This project is solely for educational purposes, and I am eager to demonstrate my skills and make a positive impact. Thank you for taking the time to review my work.