Update README.md
I am pleased to present my latest project, which focuses on the automation of work with the Chrome web browser using Selenium and Python to test end-to-end (E2E) scenarios. For the purpose of testing, I have selected a website that is devoted to aiding and preserving the lives of dogs in the United Kingdom. This project is intended solely for educational purposes, and I am excited to demonstrate my skills and make a positive impact. I appreciate your time in reviewing my work.